Number of bytes read or written should be an int, so return that

+++ chill-java/src/main/java/com/twitter/chill/SerDeState.java
@@ -50,8 +50,8 @@ public class SerDeState {
   public void setInput(byte[] in, int offset, int count) { input.setBuffer(in, offset, count); }
   public void setInput(InputStream in) { input.setInputStream(in); }
 
-  public long numOfWrittenBytes() { return output.total(); }
-  public long numOfReadBytes() { return input.total(); }
+  public int numOfWrittenBytes() { return (int)output.total(); }
+  public int numOfReadBytes() { return (int)input.total(); }
 
   // Common operations:
   public <T> T readObject(Class<T> cls) {
